indexnew & mainnew
Some parts of the website is having some problems due to our copy
pasting of code and incorrectly measuring our divs. I've re-worked the
code in the footer as well measuring out the divs. (the footer will no
longer break the page when scaling the browser.)

We may adopt this new code or we may not due to time constraints.